requirements.txt
langchain>=0.2
langchain-google-genai>=1.0
langchain-community>=0.2
langchain-text-splitters>=0.2
langchain-core>=0.2

google-generativeai>=0.5

pypdf>=4.0
sentence-transformers>=2.2

pydantic>=2.0
python-dotenv>=1.0

python-docx>=1.1

markdown>=3.6
weasyprint>=61.0
